Add simple successor generator, and checking of variables and indices
* asserts for variable limits and location indices
* throw exceptions on enum class switches default case
* Model now only contain number of variables instead of an unused vector
* initial state is now created based on the model instead of hardcoded

+++ cpp/src/State.cpp
@@ -32,7 +32,7 @@ std::ostream &operator<<(std::ostream &os, State const &m) {
             ;
 
     for (auto var : m.variables) {
-        os << std::setw(4) << var << " ";
+        os << std::setw(4) << (int)var << " ";
     }
     os << ">\n"
           "}\n"
